types cleanup
