+++ src/interfaces.ts
@@ -1,0 +1,17 @@
+export interface IItem {
+  id: string
+  optionName: string
+  pricePerUnit: number | undefined
+  quantity: number
+  category: string
+  subtotal: number | undefined
+}
+
+export interface IOption {
+  name: string
+  costPerUnit: number
+}
+
+export enum Items {
+  Materials = 'materials',
+}
